requirements.txt
boto3==1.9.207
botocore==1.12.207
docutils==0.14
jmespath==0.9.4
kafka-python==1.4.6
kazoo==2.5.0
mysql-replication==0.19
PyMySQL==0.9.3
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
tabulate==0.8.3
urllib3==1.25.3
